+++ pykechain/models/activity.py
@@ -6,7 +6,11 @@ from urllib.parse import urljoin
 
 import requests
 
-from pykechain.defaults import API_EXTRA_PARAMS, ASYNC_REFRESH_INTERVAL, ASYNC_TIMEOUT_LIMIT
+from pykechain.defaults import (
+    API_EXTRA_PARAMS,
+    ASYNC_REFRESH_INTERVAL,
+    ASYNC_TIMEOUT_LIMIT,
+)
 from pykechain.enums import (
     ActivityClassification,
     ActivityRootNames,
@@ -905,6 +909,7 @@ class Activity(TreeObject, TagsMixin):
         paper_size: PaperSize = PaperSize.A4,
         paper_orientation: PaperOrientation = PaperOrientation.PORTRAIT,
         include_appendices: bool = False,
+        include_pdf_appendices_inline: bool = True,
         include_qr_code: bool = False,
         user: Optional[User] = None,
         timeout: int = ASYNC_TIMEOUT_LIMIT,
@@ -927,6 +932,10 @@ class Activity(TreeObject, TagsMixin):
                                - landscape: landscape orientation
         :param include_appendices: True if the PDF should contain appendices, False (default)
             if otherwise.
+        :param include_pdf_appendices_inline: The PDF appendices, if appendices are enable are
+            rendered inline inside the PDF. If this is set to False, then a ZIP will be returned
+            where all attachments that cannot be rendered, including the PDF attachments are
+            included in that zip in a subdirectory 'attachments'.
         :param include_qr_code: True if the PDF should include a QR-code, False (default)
             if otherwise.
         :param user: (optional) used to calculate the offset needed to interpret Datetime
@@ -951,6 +960,9 @@ class Activity(TreeObject, TagsMixin):
                 paper_orientation, PaperOrientation, "paper_orientation"
             ),
             appendices=check_type(include_appendices, bool, "include_appendices"),
+            include_pdf_appendices_inline=check_type(
+                include_pdf_appendices_inline, bool, "include_pdf_appendices_inline"
+            ),
             includeqr=check_type(include_qr_code, bool, "include_qr_code"),
         )